chore: rename to module-config
