+++ packages/orchestration/src/util/module-config.ts
